Use runTest() with an unconfined dispatcher instead of awaitJob()
This appears to allow us to avoid explicitly waiting for coroutines in
the common case. See the comment in RobolectricTest for why this and
not StandardTestDispatcher was used.

Also opted in to the coroutines API at the module level, so we don't
have to tag every unit test with @OptIn.