renamed updateMultipleNotes and removed unused flag
Since the functions doInBackgroundUpdateMultipleNotes relies on caller to call it from background i.e it does not itself performs the operation in background, it is renamed to updateMultipleNotes.

Also the flag, shouldUpdateCards is removed because it was nowhere used. Introduced in a1fd433 and used for the first time in b9108e6 it was never called with an explicit value for this variable, so it's fine to delete it.