Change bottom navigation style to material
Bottom navigation is only used in Card types. Before, the color of
inactive tab layout and bottom navigation items was grey. The previous
removed grey from tab layout and this removes it from the bottom
navigation.

With material styling, active item has a pill-shaped indicator. This
means that we can rely on that instead of changing icon and label
colors.