[GSoC] Coroutines migration ConfSetSubdecks (#12592)
* Extracted out ConfSetSubdecks::task to confSetSubdecks

* Migrated ConfSetSubdecks to Coroutines

Since the launch is taking place from inside a AppCompatPreferenceActivity which is not a lifecycleOwner. And launchCatchingTask can only happen from a lifecycleOwner, so a custom CoroutineScope implementation is required which cancels any job (launched from this activity) when the activity gets destroyed and that is achieved by extending 'AppCompatPreferenceActivity: CoroutineScope by MainScope' and calling cancel in onDestroy.

MainScope delegates the CoroutineScope implementation with having Main dispatcher as default for launch{} calls. This is the reason why coroutine is launched using launch{} instead of launchCatchingTask{}

* Exceptions are now handled by CoroutineExceptionHandler

* Moved confSetSubdecks to DeckOptions