Adding a scoped storage migration channel (#12653)
* NF: move Channel outsile of NotificationChannels class

This will allow the next commit to requires that the NotificationChannels class
uses API 26 without restricting access to this enum. Also, `channel` is clear
enough that the full name is clear without adding `NotificationChannels` as prefix.

* NF: NotificationChannels called by Compat

Firstly, it avoid to call function when we don't need it.

Secondly, next commits will have various importances; and that depends on API
level 26, so I want to be sure to call it only in place where I've already
checked I've the right API

* NF: add a Scoped storage migration channel for notifications

* NF: scoped storage channel is lower importance

This ensure the user gets no sound notification.