Coroutine migration ChangeDeckMulti (#12700)
* Extracted ChangeDeckMulti::task to changeDeckMulti function

* ChangeDeckMulti now extends TaskDelegate instead of DismissNotes

DismissNotes is a wrapper that just executes the original task in db-transaction. This behaviour is now directly implemented in DeleteNoteMulti

* changeDeckMulti now returns actual value instead of boolean flag

* changeDeckMulti now internally executes in database transaction

* Migrated ChangeDeckMulti to Coroutines

The post change logic has been copied from ChangeDeckHandler

* Code simplification: CardBrowser

The conditional check inside the if block was mutually exclusive with the outer condition. So move the inner logic to outside. Also earlier when there was unsuccessful result, the browser would close abruptly, instead that now it displays error message

* Added todos for code cleanup and improvements

And deleted unused ChangeDeckMulti class