Create & use a call-site withQueue
`withQueue`, like `withCol`, changes the thread, and it might be more
prudent to let the logic that's closer to business to decide how and
when to change threads.

This creates a public `withQueue` that calls a block with a special,
publicly inaccessible, instance of `Queue` as a context receiver,
which makes it possible to define functions that can only be run inside
of `withQueue`.

See also:
https://github.com/ankidroid/Anki-Android/pull/12371#discussion_r990797635