Stop handling language changes manually (#12991)
* Fix test at DeckPickerTest

I don't know why, but it starts failing when the language changes are done. I'm putting this as first commit of the PR to avoid adding broken commits to the history

DeckPickerTest > [0] > com.ichi2.anki.DeckPickerTest.deckPickerOpensWithHelpMakeAnkiDroidBetterDialog[0] FAILED
    java.lang.NullPointerException
        at com.ichi2.anki.CardTemplateEditor.onSaveInstanceState(CardTemplateEditor.kt:140)
        at android.app.Activity.performSaveInstanceState(Activity.java:2037)

* Remove locale configChanges from all activities

* Remove language attachment methods

* Refactor LanguageUtil

* Upgrade language preference to use Android new app language configuration

https://developer.android.com/guide/topics/resources/app-languages

* Use the new API for changing languages on the language preference

And redirect to the system language picker if Android ≥ 13

Fix issue when setting languages with regions

e.g. pt-BR, es-AR

They didn't reload correctly. This was fixed on the 1.7x version of appcompat, so this commits bumps it to its alpha version.

note: appcompat 1.7.0alpha02 needs Kotlin 1.8+ (or at least it threw a duplicate class error while I hadn't upgraded the kotlin version)