Fix newly created deck not showing when changing card's deck in CardBrowser
The CardBrowser kept track of the available decks by checking the DeckSpinnerSelection.dropDownDecks property, this
property wasn't updating when the user would create a new deck during the process of changing a card's deck. The fix is
making CardBrowser to use directly the method that DeckSpinnerSelection was using to update dropDownDecks.